fixed import definitions after logged-user service changes
