+++ src/app/modules/logged-user/logged-user.service.ts
@@ -1,9 +1,34 @@
-import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { inject, Injectable } from '@angular/core';
+import { Observable } from 'rxjs';
+import { environment } from '../../../environments/environment';
+
+export interface AppUser {
+  id: string;
+  displayName: string;
+  phoneNumber?: string;
+  aboutMe?: string;
+  city?: string;
+}
+
+export interface UpdateAppUserDto {
+  phoneNumber?: string;
+  aboutMe?: string;
+  city?: string;
+}
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class LoggedUserService {
+  private http = inject(HttpClient);
+  private baseUrl = `${environment.apiUrl}/app-users`;
 
-  constructor() { }
+  getUser(userId: string): Observable<AppUser> {
+    return this.http.get<AppUser>(`${this.baseUrl}/${userId}`);
+  }
+
+  updateUser(userData: UpdateAppUserDto): Observable<AppUser> {
+    return this.http.patch<AppUser>(this.baseUrl, userData);
+  }
 }
